How To Play on game screen
when the player goes from the main menu to the game screen it automatically freezes their ability to move, rotate and change stone text. they have a how to play screen and buttons to go through the information/ images and once they press begin then they can move stones. On top of that whichever scenario they pick in the main menu ie, 1 or 2 will have an effect as well.